requirements.txt
asgiref==3.11.0
defusedxml==0.7.1
Django==6.0
fonttools==4.61.1
fpdf2==2.8.5
pillow==12.0.0
python-decouple==3.8
sqlparse==0.5.5
